Re-organize record/replay code to match dbt-adapters
