+++ dbt/adapters/snowflake/record/__init__.py
@@ -1,0 +1,2 @@
+from dbt.adapters.snowflake.record.cursor.cursor import SnowflakeRecordReplayCursor
+from dbt.adapters.snowflake.record.handle import SnowflakeRecordReplayHandle
